requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.4
django-js-asset==2.0.0
django-mptt==0.13.4
django-rest-framework==0.1.0
djangorestframework==3.13.1
pkg_resources==0.0.0
psycopg2-binary==2.9.3
python-dotenv==0.20.0
pytz==2022.1
sqlparse==0.4.2
